refactor(api): retire runner CLI availability module
Keep CLI persistence parsing inside runner backend health and expose a claim-level predicate so run-claim no longer depends on the legacy storage module.

X-Anneal-Run: cmtfzbrx20015mpz524iins2g
X-Anneal-Step: 1: Implementation
X-Anneal-Provider: codex
